Make zipLang a buildClient task
