+++ buildtools/tasks/client/index.js
@@ -2,7 +2,8 @@ const fs       = require("fs");
 const log      = require("fancy-log");
 const path     = require("path").posix;
 const zip      = require("gulp-zip");
-const Promise  = require("bluebird");
+const merge    = require("merge-stream");
+const rename   = require("gulp-rename");
 
 const { retryRequest } = require("../../util/downloaders.js");
 
@@ -105,11 +106,32 @@ function zipClient() {
 		.pipe(dest(DEST_FOLDER));
 }
 
+/**
+ * Zips the language files.
+ */
+ function zipLang() {
+	const resourcesPath = path.join(SHARED_DEST_FOLDER, global.OVERRIDES_FOLDER, "resources");
+
+	const opts = { nodir: true, base: resourcesPath };
+	const streams = [
+		src(path.join(resourcesPath, "pack.mcmeta"), opts),
+		src(path.join(resourcesPath, "**/*.lang"), opts)
+			.pipe(rename(f => {
+				f.dirname = path.join("assets", f.dirname);
+			}))
+	];
+
+	return merge(...streams)
+		.pipe(zip("lang.zip"))
+		.pipe(dest(DEST_FOLDER));
+}
+
 module.exports = [
 	createClientDirs,
 	saveModpackManifest,
 	copyClientOverrides,
 	copyClientLicense,
 	fetchModList,
-	zipClient
+	zipClient,
+	zipLang
 ]
